+++ otelcollector/test/ginkgo-e2e/configprocessing/suite_test.go
@@ -7,6 +7,7 @@ import (
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	promOperatorClient "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned"
 	"k8s.io/client-go/rest"
 
 	"prometheus-collector/otelcollector/test/utils"
@@ -17,7 +18,13 @@ import (
 
 var K8sClient *kubernetes.Clientset
 var Cfg *rest.Config
+var PromClient promOperatorClient.Interface
 
+/*
+ * Tests that exercise prometheus-operator CRDs (e.g. ServiceMonitors) MUST be
+ * run with the flag:
+ * -ldflags="-s -X github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring.GroupName=azmonitoring.coreos.com"
+ */
 func TestConfigProcessing(t *testing.T) {
 	RegisterFailHandler(Fail)
 
@@ -27,6 +34,8 @@ func TestConfigProcessing(t *testing.T) {
 var _ = BeforeSuite(func() {
 	var err error
 	K8sClient, Cfg, err = utils.SetupKubernetesClient()
+	Expect(err).NotTo(HaveOccurred())
+	PromClient, err = promOperatorClient.NewForConfig(Cfg)
 	Expect(err).NotTo(HaveOccurred())
 })
