+++ otelcollector/test/testkube/config-processing-test-crs/testkube-config-test-secrets-access-namespaces-crs.yaml
@@ -1,0 +1,67 @@
+apiVersion: testworkflows.testkube.io/v1
+kind: TestWorkflow
+metadata:
+  name: configprocessingcommon
+  namespace: testkube
+  labels:
+    executor: ginkgo-executor
+    test-type: ginkgo-test
+spec:
+  content:
+    git:
+      uri: https://github.com/Azure/prometheus-collector
+      revision: $BRANCH_NAME
+      paths:
+      - otelcollector/test/ginkgo-e2e
+  steps:
+    - name: Run test
+      container:
+        workingDir: /data/repo/otelcollector/test/ginkgo-e2e
+        image: cerebro31/ginkgo-test:latest
+        env:
+        - name: AMW_QUERY_ENDPOINT
+          value: "$AMW_QUERY_ENDPOINT"
+        - name: AZURE_CLIENT_ID
+          value: "$AZURE_CLIENT_ID"
+        - name: GOTOOLCHAIN
+          value: "auto"
+      shell: |
+        ginkgo run -r -p --label-filter "(config-processing-common, config-processing-common-with-config-maps)" ./configprocessing
+  pod:
+    nodeSelector:
+      kubernetes.io/os: linux
+      kubernetes.io/arch: amd64
+---
+apiVersion: testworkflows.testkube.io/v1
+kind: TestWorkflow
+metadata:
+  name: secretsaccessnamespaces
+  namespace: testkube
+  labels:
+    executor: ginkgo-executor
+    test-type: ginkgo-test
+spec:
+  content:
+    git:
+      uri: https://github.com/Azure/prometheus-collector
+      revision: $BRANCH_NAME
+      paths:
+      - otelcollector/test/ginkgo-e2e
+  steps:
+    - name: Run test
+      container:
+        workingDir: /data/repo/otelcollector/test/ginkgo-e2e
+        image: cerebro31/ginkgo-test:latest
+        env:
+        - name: AMW_QUERY_ENDPOINT
+          value: "$AMW_QUERY_ENDPOINT"
+        - name: AZURE_CLIENT_ID
+          value: "$AZURE_CLIENT_ID"
+        - name: GOTOOLCHAIN
+          value: "auto"
+      shell: |
+        ginkgo run -r --ldflags "-s -X github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring.GroupName=azmonitoring.coreos.com" --label-filter "(config-processing-secrets-access-namespaces)" ./configprocessing
+  pod:
+    nodeSelector:
+      kubernetes.io/os: linux
+      kubernetes.io/arch: amd64
